+++ forms/__init__.py
@@ -1,0 +1,13 @@
+# app/forms/__init__.py
+from .input_form import InputForm, SequenceItem
+from .constants import SimilarityAlgorithm, SmoothingType, FormConfig
+from .validators import FileValidator
+
+__all__ = [
+    'InputForm',
+    'SequenceItem',
+    'SimilarityAlgorithm',
+    'SmoothingType',
+    'FormConfig',
+    'FileValidator'
+]
